+++ SEMS/Application/SiteManagement.cs
@@ -26,6 +26,10 @@ namespace SEMS.Application
 
         public static bool deleteSite(Site deleteSite)
         {
+            if (dataHandler.getEmployeesBySiteId(deleteSite.Id).Count != 0)
+            {
+                throw new Exception("There are Employees tied to this Site");
+            }
             return dataHandler.deleteSite(deleteSite);
         }
     }
